% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcawg_load_avail.R
\name{pcawg_load}
\alias{pcawg_load}
\title{Load PCAWG mafs}
\usage{
pcawg_load(cohort)
}
\arguments{
\item{cohort}{abbreviation of PCAWG project. See pcawg_available() for valid values (string)}
}
\value{
MAF object compatible with maftools
}
\description{
Load PCAWG mafs
}
\examples{
pcawg_load("Biliary-AdenoCA")
}
